Add Viewsfolder and renamed components
