--- src/components/backupindex.vue
@@ -1,30 +1,0 @@
-<template> 
-<header class="header">
-        <a href="#" class="logoname"> Festive Bites</a>
-        
-        <nav class="navbar">
-            <ul class="navmenu">
-                <li><a href="#">Home</a></li>
-                <li><a href="#">Search</a></li>
-            </ul>
-        </nav>
-    </header>
-
-    <main class="mainContainer">
-        
-        <div class="mainText">
-            <p> 2026 Celebration</p>
-        </div>
-        <div class="mainText">
-            <h1> New Year's Eve Recepies </h1>
-        </div>
-        
-        <ul class=" categories">
-            <li>Förätter</li>
-            <li>Huvudrätter</li>
-            <li>Efterrätter</li>
-            <li>Drinkar</li>
-            
-        </ul>
-    </main>
-</template>
